backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.0
google-cloud-firestore==2.13.1
google-generativeai>=0.8.5
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
PyJWT==2.8.0
